[frontend] refactor metadata
