+++ apps/portal-front/app/(public)/layout.tsx
@@ -1,5 +1,4 @@
-import { serverFetchGraphQL } from '@/relay/serverPortalApiFetch';
-import SettingsQuery, { settingsQuery } from '@generated/settingsQuery.graphql';
+import { getDefaultMetadata } from '@/utils/generate-metadata';
 import { Button } from 'filigran-ui/servers';
 import 'filigran-ui/theme.css';
 import { Metadata } from 'next';
@@ -8,57 +7,7 @@ import LogoXTMDark from '../../public/logo_xtm_hub_dark.svg';
 import '../../styles/globals.css';
 
 export async function generateMetadata(): Promise<Metadata> {
-  const settingsResponse = await serverFetchGraphQL<settingsQuery>(
-    SettingsQuery,
-    {},
-    { cache: 'force-cache' }
-  );
-  const baseUrl = settingsResponse.data.settings.base_url_front;
-  return {
-    title:
-      'Your Gateway to Cyber Threat Intelligence & Breach & Attack Simulation | XTM Hub by Filigran',
-    description:
-      "XTM Hub is your gateway to Filigran's cybersecurity solutions: Cyber Threat Intelligence & Breach & Attack Simulation. Uncover Threats. Take Action.",
-    metadataBase: new URL(baseUrl),
-    openGraph: {
-      title: 'XTM Hub by Filigran - Cyber Threat Intelligence Platform',
-      description:
-        "XTM Hub is your gateway to Filigran's cybersecurity solutions: Cyber Threat Intelligence & Breach & Attack Simulation. Uncover Threats. Take Action.",
-      url: baseUrl,
-      siteName: 'XTM Hub by Filigran',
-      images: [
-        {
-          url: `${baseUrl}/opengraph-image.png`,
-          width: 1200,
-          height: 630,
-          alt: 'XTM Hub by Filigran',
-        },
-      ],
-      locale: 'en_US',
-      type: 'website',
-    },
-    twitter: {
-      card: 'summary_large_image',
-      title: 'XTM Hub by Filigran - Cyber Threat Intelligence Platform',
-      description:
-        "XTM Hub is your gateway to Filigran's cybersecurity solutions.",
-      images: [`${baseUrl}/opengraph-image.png`],
-      creator: '@FiligranHQ',
-      site: '@FiligranHQ',
-    },
-    alternates: {
-      canonical: baseUrl,
-    },
-    icons: {
-      icon: [
-        { url: '/favicon.ico' },
-        { url: '/favicon-16x16.png', type: 'image/png', sizes: '16x16' },
-        { url: '/favicon-32x32.png', type: 'image/png', sizes: '32x32' },
-        { url: '/favicon-96x96.png', type: 'image/png', sizes: '96x96' },
-      ],
-      apple: [{ url: '/apple-icon.png', sizes: '180x180' }],
-    },
-  };
+  return await getDefaultMetadata();
 }
 
 export default function RootLayout({
