+++ apps/portal-front/src/components/connectors/service-instance-connector-card.tsx
@@ -3,11 +3,13 @@ import { FeatureFlag } from '@/utils/constant';
 import { isFeatureEnabled } from '@/utils/settings.service';
 import { ServiceDefinitionIdentifierEnum } from '@generated/models/ServiceDefinitionIdentifier.enum';
 import { ServiceInstanceCreationStatusEnum } from '@generated/models/ServiceInstanceCreationStatus.enum';
+import { getTranslations } from 'next-intl/server';
 
 export const ServiceInstanceConnectorCard = async () => {
   const isConnectorsPageEnabled = await isFeatureEnabled(
     FeatureFlag.CONNECTORS_PAGE
   );
+  const t = await getTranslations();
   return (
     <>
       {isConnectorsPageEnabled && (
@@ -16,10 +18,9 @@ export const ServiceInstanceConnectorCard = async () => {
           serviceInstance={{
             id: 'Connector',
             creation_status: ServiceInstanceCreationStatusEnum.CREATED,
-            name: 'OpenCTI Connectors',
+            name: t('Service.Connectors.Name'),
             slug: 'opencti-connectors',
-            description:
-              'Explore a range of OpenCTI Connectors shared by the Filigran team',
+            description: t('Service.Connectors.Description'),
             illustration_document_id: '' as string,
             logo_document_id: '' as string,
             service_definition_identifier:
